fix(sidebar): address code review findings for drag-to-reorder
